Add dhcp-request-options setting
Add a setting to request additional dhcp options from a DHCP server.
This is intended to be used for external software like dispatcher
scripts, which want to query these additional option values.

As these options may contain binary data, they are serialized as a
hexadecimal string. Currently this is only implemented for the internal
DHCPv4 client.