bond: change default value of `arp_missed_max`
NM allows arp_missed_max to be in the range 0-255, while the permissible range
for arp_missed_max is 1-255 as specified by the kernel, with default value as
2. Change the same.

NM uses 0 to indicate that `arp_missed_max` is unset, however this
approach causes warnings to arise since kernel does not allow it.
Therefore change it to not apply or reapply `arp_missed_max` when it is not
supported by the bonding mode.

Signed-off-by: Pradyumn Rahar <pradyumn.rahar@oracle.com>