ovs: print clearer logs of the ovsdb "update" event
These lines says things like "changed a bridge", what seems to mean that
NM is doing the change. Actually, these logs indicate changes that NM is
being notified of, and they may even be external changes.

First, add the "monitor:" prefix to show that it's something that NM is
monitoring, not doing. Second, say "bridge changed" instead of "changed
a bridge", which sounds an action that we're doing. Finally, print the
bridge/port/iface name first, instead of the uuid-like key which is not
useful for a quick look by a human.

Before:
  ovsdb: obj[bridge:8c975244-cb0a-4add-8901-c398dcbc27d6]: changed a bridge: br-int, ...

After:
  ovsdb: monitor: br-int: changed a bridge: obj[bridge:8c975244-cb0a-4add-8901-c398dcbc27d6], ...