Implement DHCPv6 vendor option support
Add support to request and parse the DHCPv6 vendor specific option, when
it is specified in the ipv6.dhcp-request-options setting.

As the systemd implementation used by NetworkManager as a backend only
provides API calls to query explicit options from a lease, a switch or
if/else-if approach has to be used to forward all supported options.
This commit only handles the vendor options.

Also the systemd implementation already parses the value of the vendor
options. Therefore repack the parsed values into one hex string to be
passed to the dispatecher script. The format mimics the dhcp option
value, but avoids packing multiple sub options of one enterprise id and
instead repeats the enterprise id for each sub option.

Furthermore systemd also silently drops vendor specific options which
fail to parse correctly.