platform: add nexthop support
Add support for route nexthop objects to the platform. Contrary to
other object types, we don't track nexthops in the platform cache
because on some systems there could be many of them and this can
become very inefficient. To configure a desired list of nexthops for
an interface it is necessary to first dump the current list, delete
the unwanted ones and add the desired ones. The dump returns only the
nexthops for the given interface and address family.